+++ HISTORY.rst
@@ -1,14 +1,22 @@
 History
 =======
 
-2.2.3 (unreleased)
-------------------
+2.4 (unreleased)
+----------------
+
+- Nothing changed yet.
+
+
+2.3 (2025-06-10)
+----------------
 
 - Add processing algorithm to add NWRW parameters to surface parameters (#333)
 - Fix removing of cross-section locations associated to channels not modified by object import (#378)
 - Processing algorithms 'Map surfaces/DWF to connection nodes' now use point on surface instead of centroid as start vertex of mapping line (#396)
 - Reset unspecified import settings on loading new import configuration (#286)
 - Fix handling string keys in import with value_map (#279)
+- Changing Material back to what it was does not update the friction value (#345)
+- Allow digitising and importing pipes with > 2 vertices (#342)
 
 
 2.2.2 (2025-05-20)